Reorganization to Facilitate Context Injection
