+++ docstring_ai/lib/prompt_utils.py
@@ -1,477 +1,165 @@
-# Docstring generated by docstring-ai : http://github.com/ph-ausseil/docstring-ai
-"""
-This module provides functionalities to initialize and manage an AI assistant
-for adding docstrings to Python code. It utilizes OpenAI's API to create and
-interact with the assistant, manage threads, and construct prompts based on
-context from ChromaDB.
 
-Functions:
-- initialize_assistant: Initialize or retrieve an existing assistant.
-- update_assistant_tool_resources: Update the assistant's resources with file IDs.
-- create_thread: Create a new thread for the assistant's interaction.
-- construct_few_shot_prompt: Constructs a few-shot prompt using context summaries.
-- generate_few_shot_examples: Generates few-shot examples based on context.
-- extract_code_from_message: Extracts code blocks from the assistant's messages.
-"""
 
-import time
-import openai
-from openai.types.beta import vector_store_create_params
-import chromadb
-from docstring_ai.lib.chroma_utils import get_relevant_context
-import logging
-from typing import List, Dict, Callable
-from docstring_ai.lib.config import MODEL, RETRY_BACKOFF, setup_logging, MAX_RETRIES
-from pydantic import BaseModel, Field
 import json
+import os
+import logging
+from pathlib import Path
+from typing import List, Dict, Tuple, Optional
+from functools import partial
+from datetime import datetime
 
-setup_logging()
+import openai
+import tiktoken
+from tqdm import tqdm
 
-ASSISTANTS_DEFAULT_TOOLS = [
-                {"type": "code_interpreter"}, 
-                {"type": "file_search"},
-            ]
-
-
-class PythonFile(BaseModel):
-    new_file_content: str = Field(description="Updated python script with the updated docstrings.")
-
-
-def initialize_assistant(api_key: str, assistant_name: str = "DocstringAssistant") -> str:
-    """
-    Initialize or retrieve an existing Assistant.
-
-    This function checks for existing assistants by name and returns the
-    assistant ID if found. If not, it creates a new assistant with specified
-    instructions.
-
-    Args:
-        api_key (str): The API key for OpenAI authentication.
-        assistant_name (str): The name of the assistant to retrieve or create. Default is "DocstringAssistant".
-
-    Returns:
-        str: The ID of the Assistant, or None if an error occurred.
-
-    Raises:
-        Exception: If there is an error in retrieving or creating the Assistant.
-    """
-    try:
-        response = openai.beta.assistants.list()
-        for assistant in response.data:
-            if assistant.name == assistant_name:
-                logging.debug(f"Assistant '{assistant_name}' found with ID: {assistant.id}")
-                return assistant.id
-
-        instructions = (
-            "You are an AI assistant specialized in adding comprehensive docstrings to Python code. "
-            "Ensure that all functions, classes, and modules have clear and docstrings. "
-            "Docstrings should give extensive context and explain purpose, parameters, return values, and any exceptions raised."
-        )
-        assistant = openai.beta.assistants.create(
-            name=assistant_name,
-            description="Assistant to add docstrings to Python files.",
-            model=MODEL,
-            tools=ASSISTANTS_DEFAULT_TOOLS,
-            instructions=instructions
-        )
-        logging.debug(f"Assistant '{assistant_name}' created with ID: {assistant.id}")
-        return assistant.id
-    except Exception as e:
-        logging.error(f"Error initializing Assistant: {e}")
-        return None
+from docstring_ai.lib.utils import (
+    create_backup
+)
+from docstring_ai.lib.llm_utils import (
+    create_file_with_docstring,
+    generate_file_description,
+    upload_files_to_openai
+)
+from docstring_ai.lib.chroma_utils import embed_and_store_files
+from pydantic import BaseModel, Field
 
 
-def update_assistant_tool_resources(api_key: str, assistant_id: str, file_ids: List[str]) -> None:
-    """
-    Update the Assistant's tool resources with the uploaded file IDs.
 
-    This function creates a vector store for the files and updates the assistant's
-    tool resources to include the new vector store.
-
-    Args:
-        api_key (str): The API key for OpenAI authentication.
-        assistant_id (str): The ID of the assistant to update.
-        file_ids (List[str]): A list of file IDs to add to the assistant's resources.
-
-    Raises:
-        Exception: If there is an error while updating the assistant's resources.
-    """
-    try:
-        vector_store_id = create_vector_store(f"Docstring-AI::{assistant_id}", file_ids)
-        openai.beta.assistants.update(
-            assistant_id=assistant_id,
-            tool_resources={
-                "file_search": {
-                    "vector_store_ids": [vector_store_id]
-                }
-            }
-        )
-        logging.debug(f"Assistant '{assistant_id}' tool_resources updated with {len(file_ids)} files.")
-    except Exception as e:
-        logging.error(f"Error updating Assistant's tool_resources: {e}")
+class TreeConfig(BaseModel):
+    base_path: Path = Field(description="Base Path of the project")
+    repositories_to_ignore: List[str] = Field(default = [] , description="List of repository paths to ignore.")
+    extensions_to_ignore: List[str] = Field(default = ['.daicache'], description="List of file extensions to ignore.")
+    apply_gitignore_policy: Optional[bool] = Field(default=True, description="Whether to apply .gitignore rules.")
 
 
-def create_thread(api_key: str, assistant_id: str, initial_messages: List[dict] = None) -> str:
-    """
-    Create a new Thread for the Assistant.
+def dump_tree(path: str, config: TreeConfig, prefix: str = ""):
+    path = Path(path)
 
-    A thread is used for maintaining a conversation context with the assistant.
+    # Compute the relative path for comparison with repositories_to_ignore
+    relative_path = str(path.relative_to(config.base_path))
+    # Check if the current path is in the repositories to ignore
+    if relative_path in [str(Path(ignored)) for ignored in config.repositories_to_ignore]:
+        return ""  # Skip this repository
+    # Initialize the tree string representation
+    tree_str_representation = f"{prefix}{path.name}/\n"
 
-    Args:
-        api_key (str): The API key for OpenAI authentication.
-        assistant_id (str): The ID of the assistant for which to create a thread.
-        initial_messages (List[dict], optional): A list of initial messages to start the thread.
+    # Load .gitignore rules if applicable
+    gitignore_patterns = []
+    if config.apply_gitignore_policy and (path / ".gitignore").exists():
+        with open(path / ".gitignore") as f:
+            gitignore_patterns = [line.strip() for line in f if line.strip() and not line.startswith("#")]
 
-    Returns:
-        str: The ID of the created thread, or None if an error occurred.
+    for p in path.iterdir():
+        # Skip directories or files that match .gitignore or configuration rules
+        if any(p.match(pattern) for pattern in gitignore_patterns):
+            continue
+        if p.is_dir() and str(p.resolve()) in [str(Path(config.base_path).joinpath(ignored).resolve()) for ignored in config.repositories_to_ignore]:
+            continue
+        if p.is_file() and any(p.suffix == ext for ext in config.extensions_to_ignore):
+            continue
 
-    Raises:
-        Exception: If there is an error creating the thread.
-    """
-    try:
-        payload = {
-            "messages": initial_messages if initial_messages else []
-        }
-        thread = openai.beta.threads.create(**payload)
-        logging.debug(f"Thread created with ID: {thread.id}")
-        return thread.id
-    except Exception as e:
-        logging.error(f"Error creating Thread: {e}")
-        return None
+        # Recursive call for directories or add file to the tree
+        if p.is_dir():
+            tree_str_representation += dump_tree(p, prefix=prefix + "  ", config=config)
+        else:
+            tree_str_representation += f"{prefix}  {p.name}\n"
 
-
-def construct_few_shot_prompt(
-    collection: chromadb.Collection,
-    classes: Dict[str, List[str]],
-    max_tokens: int,
-    context: str = None
-    ) -> str:
-    """
-    Constructs a few-shot prompt using context summaries from ChromaDB.
-
-    This function retrieves relevant context based on class dependencies and
-    constructs a prompt for the assistant.
-
-    Args:
-        collection (chromadb.Collection): The ChromaDB collection to query for context.
-        classes (Dict[str, List[str]]): A dictionary containing class names and their parent classes.
-        max_tokens (int): The maximum number of tokens to be used in the prompt.
-
-    Returns:
-        str: The constructed few-shot prompt.
-
-    Raises:
-        Exception: If there is an error retrieving context or generating the prompt.
-    """
-    try:
-        documents = get_relevant_context(collection=collection,
-        classes=classes,
-        max_tokens=max_tokens // 2,
-        where={"file_type": "script"},
-        )
-
-        context = ""
-
-        if documents:    
-            context += "Python classes with comprehensive docstrings:\n\n"
-            context +=f"{documents}\n\n"
-        if context: 
-            context +="More informations\n"
-            context +=f"{context}\n\n"
-        
-        return context
-    except Exception as e:
-        logging.error(f"Error constructing few-shot prompt: {e}")
-        return ""
+    return tree_str_representation
 
 
-def send_message_to_assistant(
+
+def generate_folder_descriptions(repo_path: Path, file_tree: str, generate_readme: bool = False, manual: bool = True):
+    logging.debug("To be implemented : This function will create a summary of what is contained in a specific Folder")
+    return {}
+
+def generate_files_descriptions(
+    files_to_describe: List[str],
+    output_dir: Path,
     assistant_id: str,
     thread_id: str,
-    prompt: str,
-    response_format: BaseModel = None,
-    tools: List = [],
-    tool_choice = "auto",
-    functions: Dict[str, Callable] = {}
-) -> str:
-    """
-    Sends a prompt to the Assistant and retrieves the response.
-
-    Args:
-        assistant_id (str): The ID of the Assistant.
-        thread_id (str): The ID of the thread for communication.
-        prompt (str): The prompt or content to send to the Assistant.
-
-    Returns:
-        str: The Assistant's response text, or an error message if an issue occurs.
-    """
-    try:
-        message = openai.beta.threads.messages.create(
-            thread_id=thread_id,
-            role="user",
-            content=prompt,
-        )
-
-        run = openai.beta.threads.runs.create(
-            thread_id=thread_id,
-            assistant_id=assistant_id,
-            response_format=response_format,
-            tool_choice=tool_choice,
-            tools=ASSISTANTS_DEFAULT_TOOLS + tools
-        )
-        if poll_run_completion(
-            run_id=run.id,
-            thread_id=thread_id,
-            functions=functions
-        ):
-            last_assistant_message = retrieve_last_assistant_message(thread_id)
-            return last_assistant_message[-1].text.value
-        return "Operation failed due to incomplete run."
-    except IndexError as e:
-        print(f"last_assistant_message : {last_assistant_message}")
-        raise e
-    except Exception as e:
-        print(f"Response format is : {response_format}")
-        logging.error(f"Error during interaction with Assistant: {e}")
-        return f"Operation failed due to an API error {e}."
-
-
-def generate_file_description(assistant_id: str, thread_id: str, file_content: str) -> str:
-    """
-    Generates a detailed description of a Python file using the Assistant.
-
-    Args:
-        assistant_id (str): The ID of the Assistant.
-        thread_id (str): The ID of the thread for communication.
-        file_content (str): The content of the Python file.
-
-    Returns:
-        str: A detailed description of the file.
-    """
-    prompt = (
-        "Provide a comprehensive & detailed description of the following Python file. "
-        "Highlight its main functionalities, purpose, classes, and function constructors. "
-        "Include any important details that would help understand the purpose, functionality, context, structure, and intent of the code.\n\n"
-        "```python\n"
-        f"{file_content}\n"
-        "```"
-    )
-    return send_message_to_assistant(
-        assistant_id=assistant_id, thread_id=thread_id, prompt=prompt)
-
-
-def create_file_with_docstring(
-    assistant_id: str,
-    thread_id: str,
-    code: str,
-    context: str,
-    functions: Dict[str, Callable]
-) -> str:
-    """
-    Adds docstrings to Python code using the Assistant.
-
-    Args:
-        assistant_id (str): The ID of the Assistant.
-        thread_id (str): The ID of the thread for communication.
-        code (str): The Python code to process.
-        context (str): Contextual examples or instructions for generating docstrings.
-
-    Returns:
-        str: The code with added docstrings, or None if an error occurs.
-    """
-
-    initial_prompt = "You will be asked to generate dosctrings for a Python . To do so we will give you some contextual information in the section (Context), then the script in the section (Instructions) .\n"
-
-    escaped_code = code.replace('` ``', '`  ``')
-    escaped_code = code.replace('```', '` ``')
-    instructions = (
-        "Please add appropriate docstrings to the following Python code. "
-        "Ensure that all functions, classes, and modules have clear and concise docstrings explaining their purpose, parameters, return values, and any exceptions raised.\n\n"
-        "### Original python script :"
-        "```python\n"
-        f"{escaped_code}\n"
-        "```"
-    )
-    if not context:
-        context = "We haven't been able to provide additional context"
-    
-    final_prompt = initial_prompt 
-    final_prompt += f"\n\n## Context\n\n{context}"
-    final_prompt += f"\n\n## Instructions\n\n{instructions}"
-
-
-    try:
-        response = send_message_to_assistant(
-            assistant_id=assistant_id,
-            thread_id=thread_id,
-            prompt=final_prompt,
-            tool_choice={"type": "function", "function": {"name": "write_file_with_new_docstring"},
-            tools=[
-                {"type": "function",
-                    "function": {
-                        "name": "write_file_with_new_docstring",
-                        "description": "Add docstrings to a python file (.py).",
-                        "parameters": {
-                            "type": "object",
-                            "properties": {
-                                "new_file_content": {
-                                    "type": "string",
-                                    "description": "Content of the python file (.py). This content should be the original script + added docstrings."
-                                }
-                            },
-                            "required": ["new_file_content"]
-                        }
-                    }
-                }
-            ],
-            functions=functions
-        )
-    except Exception as e: 
-        print(f"Error : {e}")
-        raise Exception(e)
-
-    if response:
-        logging.debug(f"The Response is : {response[:200]}")
-        return response
-    return None
-
-
-# Utility Functions
-
-def create_vector_store(vector_store_name: str, file_ids: List[str]) -> str:
-    """
-    Creates a vector store and associates it with file IDs.
-
-    Args:
-        vector_store_name (str): Name for the vector store.
-        file_ids (List[str]): List of file IDs to associate with the vector store.
-
-    Returns:
-        str: The ID of the created vector store.
-    """
-    vector_store = openai.beta.vector_stores.create(
-        name=vector_store_name,
-        expires_after=vector_store_create_params.ExpiresAfter(
-            anchor="last_active_at",
-            days=3
-        )
-    )
-    openai.beta.vector_stores.file_batches.create(
-        vector_store_id=vector_store.id,
-        file_ids=file_ids,
-    )
-    return vector_store.id
-
-
-def poll_run_completion(
-    run_id: str,
-    thread_id: str,
-    functions: Dict[str, Callable]
-) -> bool:
-    """
-    Polls until the run is completed, failed, or cancelled, with a retry mechanism.
-
-    Args:
-        run_id (str): The ID of the run to monitor.
-        thread_id (str): The thread ID associated with the run.
-
-    Returns:
-        bool: True if the run completed successfully, False otherwise.
-    """
-    retries = 0
-
-    while retries <= MAX_RETRIES:
-        while True:
+    context_summary: List[Dict],
+    collection,
+    api_key: str,
+    repo_path: str,
+    project_tree: str,
+    directory_descriptions: Dict[str, str]
+    ):
+    file_descriptions_list = []
+    for file in files_to_describe:
+        relative_path = str(Path(os.path.relpath(file, repo_path)))
+        if not any(str(Path(entry["file"])) == relative_path for entry in context_summary):
             try:
-                current_run = openai.beta.threads.runs.retrieve(
-                    run_id=run_id,
-                    thread_id=thread_id
+
+
+                file_description = generate_file_description(
+                    assistant_id=assistant_id,
+                    thread_id=thread_id,
+                    project_tree=project_tree,
+                    directory_descriptions=directory_descriptions,
+                    file_path=file
                 )
-                status = current_run.status
-                # Log status for debugging
-                logging.debug(f"Run {run_id} current status: {status}")
-                
-                if status == 'completed':
-                    logging.debug(f"Run {run_id} completed.")
-                    # Ensure the thread has at least one assistant message
-                    last_message = retrieve_last_assistant_message(thread_id)
-                    if last_message:
-                        return True
-                    logging.error("Run completed, but no assistant response available.")
-                    return False
-                elif status in ['failed', 'expired', 'cancelled']: 
-                    logging.error(f"Run {run_id} ended with status: {status}")
-                    logging.error(f"Details : {current_run.last_error}")
-                    retries += 1
-                    time.sleep(RETRY_BACKOFF)
-                    return False
-                else:
-                    logging.debug(f"Run {run_id} still in progress. Waiting...")
-                    if status == "requires_action":
-                        try : 
-                            for tool_call in current_run.required_action.submit_tool_outputs.tool_calls:
-                                if tool_call.function.name == "write_file_with_new_docstring":
-                                    return_value = functions[tool_call.function.name](**json.loads(tool_call.function.arguments))
 
-                                    if return_value:
-                                        openai.beta.threads.runs.submit_tool_outputs(
-                                            thread_id=thread_id,
-                                            run_id=run_id,
-                                            tool_outputs=[
-                                                {
-                                                    "tool_call_id": tool_call.id,
-                                                    "output": ""
-                                                }
-                                            ]
-                                        )
-                                logging.debug(f"Tool called : {tool_call.function.name}")
-                                logging.debug(f"Tool returned : {str(return_value)}")
-                        except Exception as e: 
-                            logging.error(f"Exception raised during Tool Call: {e}")
-                            retries += 1
-                            time.sleep(RETRY_BACKOFF)
-                            return False
-                        
-                    time.sleep(RETRY_BACKOFF)
+                # Save description
+                description_file_path = output_dir / Path(file).with_suffix('.txt')
+                description_file_path.parent.mkdir(parents=True, exist_ok=True)
+                with open(description_file_path, 'w', encoding='utf-8') as f:
+                    f.write(file_description)
+
+                file_descriptions_list.append(str(description_file_path))
+                context_summary.append({"file": relative_path, "description": file_description})
+
             except Exception as e:
-                logging.error(f"An error occurred while polling the run: {e}")
-                break  # Exit the inner loop to retry
-        retries += 1
-        if retries > MAX_RETRIES:
-            logging.error(f"Maximum retries reached for run {run_id}. Aborting.")
-            return False
-        
-        logging.debug(f"Retrying run {run_id} (attempt {retries}/{MAX_RETRIES})...")
-        time.sleep(RETRY_BACKOFF * retries)  # Exponential backoff
+                logging.error(f"Failed to generate description for {file}: {e}")
+    return file_descriptions_list
 
-    return False
-
-
-def retrieve_last_assistant_message(thread_id: str) -> str:
+def generate_descriptions(
+    files_to_describe: List[str],
+    output_dir: Path,
+    assistant_id: str,
+    thread_id: str,
+    context_summary: List[Dict],
+    collection,
+    api_key: str,
+    repo_path: Path  # Add repo_path parameter
+) -> List[str]:
     """
-    Retrieves the last message from a thread.
+    Generates detailed descriptions for files, embeds them into ChromaDB, and uploads to OpenAI, updating the Assistant's resources.
 
     Args:
-        thread_id (str): The thread ID from which to retrieve the message.
+        files_to_process (List[str]): List of file paths to generate descriptions for.
+        output_dir (Path): Directory to store description files.
+        assistant_id (str): OpenAI Assistant ID.
+        thread_id (str): OpenAI Thread ID.
+        context_summary (List[Dict]): Current context summary.
+        collection: ChromaDB collection.
+        api_key (str): OpenAI API key.
+        repo_path (str): Repository path for computing relative paths.
 
     Returns:
-        str: The last message content, or None if no message is found.
+        List[str]: List of successfully uploaded description file IDs.
     """
-    thread_messages = openai.beta.threads.messages.list(
-        thread_id=thread_id,
-        order="asc"
-    ).data
-    if not thread_messages:
-        logging.error(f"No messages found in Thread: {thread_id}")
-        return None
+    description_file_ids = []
+    repo_path = Path(repo_path)
+    tree_config = TreeConfig(base_path= repo_path)
+    project_tree = dump_tree(path= repo_path, config= tree_config)
 
-    if hasattr(thread_messages[-1],'role') : 
-        logging.debug(f"##### Success : ")
-        logging.debug(f"role:{thread_messages[-1].role}")
-        logging.debug(f"create_at:{thread_messages[-1].created_at}")
-        logging.debug(f"status:{thread_messages[-1].status}\n")
-    else :
-        logging.error(f"##### Failure : ")
-        logging.error(thread_messages)
+    directory_descriptions = generate_folder_descriptions(repo_path =repo_path, file_tree = project_tree)
 
-    return thread_messages[-1].content
+    file_descriptions_list = generate_files_descriptions(
+            files_to_describe=files_to_describe,
+            output_dir=output_dir,
+            assistant_id=assistant_id,
+            thread_id=thread_id,
+            context_summary=context_summary,
+            collection=collection,
+            api_key=api_key,
+            repo_path=repo_path,
+            project_tree=project_tree,
+            directory_descriptions=directory_descriptions
+        )
+    # Embed and upload descriptions
+    if file_descriptions_list:
+        embed_and_store_files(collection, file_descriptions_list, tags={"file_type": "description"})
+        description_file_ids = upload_files_to_openai(file_descriptions_list)
+
+    return description_file_ids
